LPD-18488 Isolate useOnToggleSidebars to do better testing
