+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/thunks/switchSidebarPanel.tsx
@@ -7,6 +7,7 @@ import {switchSidebarPanel as switchSidebarPanelAction} from '../actions/index';
 import hideProductMenuIfPresent from '../utils/hideProductMenuIfPresent';
 
 interface Action {
+	hidden?: boolean;
 	itemConfigurationOpen?: boolean;
 }
