+++ modules/apps/layout/layout-content-page-editor-web/types/src/main/resources/META-INF/resources/page_editor/app/thunks/switchSidebarPanel.d.ts
@@ -5,6 +5,7 @@
 
 import {switchSidebarPanel as switchSidebarPanelAction} from '../actions/index';
 interface Action {
+	hidden?: boolean;
 	itemConfigurationOpen?: boolean;
 }
 export default function switchSidebarPanel(
